app/requirements.txt
Flask==3.0.0
azure-storage-blob==12.19.0
gunicorn==21.2.0